Fix 64-bit compilation errors.  Code only passes semantic analysis, hasn't been unit tested in 64-bit environment yet.
git-svn-id: http://svn.dsource.org/projects/dstats/trunk@226 2b181080-74b1-4833-b132-872db0b9d729

+++ regress.d
@@ -946,7 +946,7 @@ bool[] toBools(R)(R range) {
     return tempdup(map!"(a) ? true : false"(range));
 }
 
-auto toRandomAccessRoR(T)(uint len, T ror) {
+auto toRandomAccessRoR(T)(size_t len, T ror) {
     static assert(isRoR!T);
     alias ElementType!T E;
     static if(isRandomAccessRange!T && isRandomAccessRange!E) {
